Upload code

+++ MF1671_BHTuyen/aspnetcore/MISA.WebFresher052023/MISA.WebFresher052023.Application/Service/TransferAssetService.cs
@@ -87,23 +87,19 @@ namespace MISA.WebFresher052023.Application.Service
         }
 
         /// <summary>
-        /// Tạo mới một chứng từ
+        /// Tạo mới một chứng từ trong bảng transfer_asset
         /// </summary>
-        /// <param name="transferAssetDto">Chứng từ tạo mới</param>
+        /// <param name="transferAsset">Chứng từ tạo mới</param>
         /// <param name="TransferAssetId">Mã Id của chứng từ</param>
         /// <returns></returns>
         /// Created By: Bùi Huy Tuyền
-        public async Task CreateAsync(TransferAssetDto transferAssetDto, Guid TransferAssetId)
+        public async Task CreateAsync(TransferAssetDto transferAsset, Guid TransferAssetId)
         {
-            _transferAssetManager.CheckTransferDateLaterTransTransactionDateAsync(transferAssetDto.TransferDate, transferAssetDto.TransactionDate);
-            // Check trùng mã code
-            await _transferAssetManager.CheckCodeConflictAsync(transferAssetDto.TransferAssetCode);
+            var transferAssetCreate = _mapper.Map<TransferAssetEntity>(transferAsset);
 
-            var transferAssetEntity = _mapper.Map<TransferAssetEntity>(transferAssetDto);
+            transferAssetCreate.SetKey(TransferAssetId);
 
-            transferAssetEntity.SetKey(TransferAssetId);
-
-            if (transferAssetEntity is BaseAuditEntity baseAuditEntity)
+            if (transferAssetCreate is BaseAuditEntity baseAuditEntity)
             {
                 baseAuditEntity.CreatedDate = DateTime.Now;
                 baseAuditEntity.CreatedBy = VietNamese.Admin;
@@ -111,7 +107,7 @@ namespace MISA.WebFresher052023.Application.Service
                 baseAuditEntity.ModifiedBy = VietNamese.Admin;
             }
 
-            await _transferAssetRepository.CreateAsync(transferAssetEntity);
+            await _transferAssetRepository.CreateAsync(transferAssetCreate);
         }
 
         /// <summary>
@@ -122,139 +118,192 @@ namespace MISA.WebFresher052023.Application.Service
         /// Created By: Bùi Huy Tuyền (10/09/2023)
         public async Task CreateTransferAssetFullAsync(TransferAssetCreateFull transferAssetCreateFull)
         {
+            // Chứng từ cần thêm mới
+            var transferAsset = transferAssetCreateFull.TransferAsset;
+
+            // Check trùng mã code chứng từ
+            await _transferAssetManager.CheckCodeConflictAsync(transferAsset.TransferAssetCode);
+
+            // Danh sách tài sản điều chuyển
+            var transferAssetDetailCreates = transferAssetCreateFull.TransferAssetDetails;
+
+            // Danh sách Id tài sản điều chuyển cần thêm
+            var fixedAssetIds = transferAssetDetailCreates.Select(transferAssetDetailCreate => transferAssetDetailCreate.FixedAssetId).ToList();
+
+            // Nếu danh sách id rỗng thì gửi thông báo lỗi cho client
+            if (fixedAssetIds == null || fixedAssetIds.Count == 0)
+            {
+                // Nếu danh sách rỗng thì gửi thông báo lỗi cho client
+                throw new UserException(VietNamese.TransferAssetDetailNotAdd);
+            }
+
+            // Kiểm tra ngày điều chuyển
+            await _transferAssetManager.CheckTransferDateAsync(fixedAssetIds, transferAsset.TransferDate, transferAsset.TransactionDate, null, ActionMode.CREATE);
+
+            // Danh sách ban giao nhận (Có thể để trống)
+            var receiverCreates = transferAssetCreateFull.Receivers;
+
+            // Tạo mới một Mã Id của chứng từ
+            var transferAssetId = Guid.NewGuid();
+
+            // Mở Transaction
             await _unitOfWork.BeginTransactionAsync();
 
             try
             {
-                // Tạo mới một Mã Id của chứng từ
-                var transferAssetId = Guid.NewGuid();
+                // Thêm chứng từ vào bảng transfer_asset
+                await CreateAsync(transferAsset, transferAssetId);
 
-                var transferAssetCreateDto = transferAssetCreateFull.TransferAsset;
-
-                await CreateAsync(transferAssetCreateDto, transferAssetId);
-
-                // Danh sách tài sản điều chuyển
-                var transferAssetDetailCreateDtos = transferAssetCreateFull.TransferAssetDetails;
-
-
-                if (transferAssetDetailCreateDtos.Any())
+                // Thêm danh sách tài sản điều chuyển
+                if (transferAssetDetailCreates != null && transferAssetDetailCreates.Any())
                 {
-                    await _transferAssetDetailService.CreateManyAsync(transferAssetDetailCreateDtos, transferAssetId);
-                }
-                else
-                {
-                    // Nếu danh sách rỗng thì gửi thông báo lỗi cho client
-                    throw new UserException(VietNamese.TransferAssetDetailNotAdd);
+                    await _transferAssetDetailService.CreateManyAsync(transferAssetDetailCreates, transferAssetId);
                 }
 
-                // Danh sách ban giao nhận (Có thể để trống)
-                var receiverCreateDtos = transferAssetCreateFull.Receivers;
-
-                if (receiverCreateDtos != null && receiverCreateDtos.Any())
+                // Thêm danh sách ban giao nhận
+                if (receiverCreates != null && receiverCreates.Any())
                 {
-                    await _receiverService.CreateManyAsync(receiverCreateDtos, transferAssetId);
+                    await _receiverService.CreateManyAsync(receiverCreates, transferAssetId);
                 }
 
+                // Commit
                 await _unitOfWork.CommitAsync();
             }
             catch (Exception)
             {
+                // Rollback khi có lỗi xảy ra
                 await _unitOfWork.RollbackAsync();
                 throw;
             }
         }
 
-        public async Task UpdateAsync(TransferAssetDto transferAssetDto)
+        /// <summary>
+        /// Cập nhật một chứng từ trong bảng transfer_asset
+        /// </summary>
+        /// <param name="transferAsset">Chứng từ cần cập nhật vào bảng transfer_asset</param>
+        /// <returns></returns>
+        public async Task UpdateAsync(TransferAssetDto transferAsset)
         {
-            _transferAssetManager.CheckTransferDateLaterTransTransactionDateAsync(transferAssetDto.TransferDate, transferAssetDto.TransactionDate);
+            var transferAssetUpdate = _mapper.Map<TransferAssetEntity>(transferAsset);
 
-            var transferAssetEntity = await _transferAssetRepository.GetAsync(transferAssetDto.TransferAssetId);
-
-            if (transferAssetEntity.GetCode() != transferAssetDto.TransferAssetCode)
-            {
-                await _transferAssetManager.CheckCodeConflictAsync(transferAssetDto.TransferAssetCode);
-            }
-
-            var transferAssetUpdateEntity = _mapper.Map<TransferAssetEntity>(transferAssetDto);
-
-            if (transferAssetUpdateEntity is BaseAuditEntity baseAuditEntity)
+            if (transferAssetUpdate is BaseAuditEntity baseAuditEntity)
             {
                 baseAuditEntity.ModifiedBy = VietNamese.Admin;
                 baseAuditEntity.ModifiedDate = DateTime.Now;
             }
 
-            await _transferAssetRepository.UpdateAsync(transferAssetUpdateEntity);
+            await _transferAssetRepository.UpdateAsync(transferAssetUpdate);
         }
 
+        /// <summary>
+        /// Cập nhật một chứng từ đầy đủ
+        /// </summary>
+        /// <param name="transferAssetUpdateFull">Chứng từ đầy đủ</param>
+        /// <returns></returns>
+        /// Created By: Bùi Huy Tuyền (10/09/2023)
         public async Task UpdateTransferAssetFullAsync(TransferAssetUpdateFull transferAssetUpdateFull)
         {
+            // Chứng từ cần cập nhật
+            var transferAsset = transferAssetUpdateFull.TransferAsset;
+
+            // Lấy ra chứng từ cần cập nhật trong database để kiểm tra
+            var transferAssetEntity = await _transferAssetRepository.GetAsync(transferAsset.TransferAssetId);
+
+            // Kiểm tra trùng mã
+            if (transferAssetEntity.GetCode() != transferAsset.TransferAssetCode)
+            {
+                await _transferAssetManager.CheckCodeConflictAsync(transferAsset.TransferAssetCode);
+            }
+
+            // Danh sách tài sản điều chuyển cần cập nhật (thêm, sửa, xóa nhiều)
+            var transferAssetDetailFlags = transferAssetUpdateFull.TransferAssetDetails;
+
+            // Danh sách tài sản điều chuyển thêm mới
+            var transferAssetDetailCreate = transferAssetDetailFlags?.Where(transferAssetDetailFlag => transferAssetDetailFlag.ActionMode == ActionMode.CREATE).Select(transferAssetDetailFlag => transferAssetDetailFlag.TransferAssetDetail);
+
+            // Danh sách tài sản điều chuyển cần cập nhật
+            var transferAssetDetailUpdate = transferAssetDetailFlags?.Where(transferAssetDetailFlag => transferAssetDetailFlag.ActionMode == ActionMode.UPDATE).Select(transferAssetDetailFlag => transferAssetDetailFlag.TransferAssetDetail);
+
+            // Danh sách Id tài sản điều chuyển cần xóa
+            var transferAssetDetailDeleteIds = transferAssetDetailFlags?.Where(transferAssetDetailFlag => transferAssetDetailFlag.ActionMode == ActionMode.DELETE).Select(transferAssetDetailFlag => transferAssetDetailFlag.TransferAssetDetail.TransferAssetDetailId).ToList();
+
+            // Danh sách Id tài sản điều chuyển cần thêm
+            var fixedAssetCreateIds = transferAssetDetailCreate?.Select(transferAssetDetailCreateDto => transferAssetDetailCreateDto.FixedAssetId).ToList();
+
+            // Danh sách Id tài sản điều chuyển cần cập nhật
+            var fixedAssetUpdateIds = transferAssetDetailUpdate?.Select(transferAssetDetailCreateDto => transferAssetDetailCreateDto.FixedAssetId).ToList();
+
+            // Danh sách Id để check
+            var fixedAssetIds = fixedAssetUpdateIds == null ? fixedAssetCreateIds : fixedAssetCreateIds?.Concat(fixedAssetUpdateIds).ToList();
+
+            if (fixedAssetIds != null)
+            {
+                // Kiểm tra ngày chứng từ cập nhật
+                await _transferAssetManager.CheckTransferDateAsync(fixedAssetIds, transferAsset.TransferDate, transferAsset.TransactionDate, transferAsset.TransferAssetId, ActionMode.UPDATE);
+            }
+
+            // Danh sách người trong ban giao nhận có gắn cờ thêm, sửa, xóa
+            var receiverFlags = transferAssetUpdateFull.Receivers;
+
+            // Danh sách tạo
+            var receiverCreate = receiverFlags?.Where(receiverFlag => receiverFlag.ActionMode == ActionMode.CREATE).Select(receiverFlag => receiverFlag.Receiver);
+
+            // Danh sách cập nhật
+            var receiverUpdate = receiverFlags?.Where(receiverFlag => receiverFlag.ActionMode == ActionMode.UPDATE).Select(receiverFlag => receiverFlag.Receiver);
+
+            // Danh sách mã ID cần xóa
+            var receiverDeleteIds = receiverFlags?.Where(receiverFlag => receiverFlag.ActionMode == ActionMode.DELETE).Select(receiverFlag => receiverFlag.Receiver.ReceiverId).ToList();
+
+            // Mở Transaction
             await _unitOfWork.BeginTransactionAsync();
 
             try
             {
-                var transferAssetUpdateDto = transferAssetUpdateFull.TransferAsset;
+                // Cập nhật chứng từ
+                await UpdateAsync(transferAsset);
 
-                await UpdateAsync(transferAssetUpdateDto);
-
-                // Thêm sửa xóa nhiều tài sản điều chuyển
-                var transferAssetDetailFlags = transferAssetUpdateFull.TransferAssetDetails;
-
-                if (transferAssetDetailFlags != null)
+                // Thêm nhiều tài sản
+                if (transferAssetDetailCreate != null && transferAssetDetailCreate.Any())
                 {
-                    var transferAssetDetailCreateDtos = transferAssetDetailFlags.Where(transferAssetDetailFlag => transferAssetDetailFlag.ActionMode == ActionMode.CREATE).Select(transferAssetDetailFlag => transferAssetDetailFlag.TransferAssetDetail);
-
-                    var transferAssetDetailUpdateDtos = transferAssetDetailFlags.Where(transferAssetDetailFlag => transferAssetDetailFlag.ActionMode == ActionMode.UPDATE).Select(transferAssetDetailFlag => transferAssetDetailFlag.TransferAssetDetail);
-
-                    var transferAssetDetailDeleteIds = transferAssetDetailFlags.Where(transferAssetDetailFlag => transferAssetDetailFlag.ActionMode == ActionMode.DELETE).Select(transferAssetDetailFlag => transferAssetDetailFlag.TransferAssetDetail.TransferAssetDetailId).ToList();
-
-                    if (transferAssetDetailCreateDtos.Any())
-                    {
-                        await _transferAssetDetailService.CreateManyAsync(transferAssetDetailCreateDtos, transferAssetUpdateDto.TransferAssetId);
-                    }
-
-                    if (transferAssetDetailUpdateDtos.Any())
-                    {
-                        await _transferAssetDetailService.UpdateManyAsync(transferAssetDetailUpdateDtos);
-                    }
-
-                    if (transferAssetDetailDeleteIds.Any())
-                    {
-                        await _transferAssetDetailService.DeleteManyAsync(transferAssetDetailDeleteIds);
-                    }
+                    await _transferAssetDetailService.CreateManyAsync(transferAssetDetailCreate, transferAsset.TransferAssetId);
                 }
 
-                // Thêm sửa xóa nhiều ban giao nhận tài sản
-                var receiverFlags = transferAssetUpdateFull.Receivers;
-
-                if (receiverFlags != null)
+                // Cập nhật nhiều tài sản
+                if (transferAssetDetailUpdate != null && transferAssetDetailUpdate.Any())
                 {
-                    var receiverCreateDtos = receiverFlags.Where(receiverFlag => receiverFlag.ActionMode == ActionMode.CREATE).Select(receiverFlag => receiverFlag.Receiver);
-
-                    var receiverUpdateDtos = receiverFlags.Where(receiverFlag => receiverFlag.ActionMode == ActionMode.UPDATE).Select(receiverFlag => receiverFlag.Receiver);
-
-                    var receiverDeleteIds = receiverFlags.Where(receiverFlag => receiverFlag.ActionMode == ActionMode.DELETE).Select(receiverFlag => receiverFlag.Receiver.ReceiverId).ToList();
-
-
-                    if (receiverCreateDtos.Any())
-                    {
-                        await _receiverService.CreateManyAsync(receiverCreateDtos, transferAssetUpdateDto.TransferAssetId);
-                    }
-
-                    if (receiverUpdateDtos.Any())
-                    {
-                        await _receiverService.UpdateManyAsync(receiverUpdateDtos);
-                    }
-
-                    if (receiverDeleteIds.Any())
-                    {
-                        await _receiverService.DeleteManyAsync(receiverDeleteIds);
-                    }
+                    await _transferAssetDetailService.UpdateManyAsync(transferAssetDetailUpdate);
                 }
 
+                // Xóa nhiều tài sản
+                if (transferAssetDetailDeleteIds != null && transferAssetDetailDeleteIds.Any())
+                {
+                    await _transferAssetDetailService.DeleteManyAsync(transferAssetDetailDeleteIds);
+                }
+
+                // Thêm nhiều ban giao nhận
+                if (receiverCreate != null && receiverCreate.Any())
+                {
+                    await _receiverService.CreateManyAsync(receiverCreate, transferAsset.TransferAssetId);
+                }
+
+                // Cập nhật nhiều ban giao nhận
+                if (receiverUpdate != null && receiverUpdate.Any())
+                {
+                    await _receiverService.UpdateManyAsync(receiverUpdate);
+                }
+
+                // Xóa nhiều ban giao nhận
+                if (receiverDeleteIds != null && receiverDeleteIds.Any())
+                {
+                    await _receiverService.DeleteManyAsync(receiverDeleteIds);
+                }
+
+                // Commit
                 await _unitOfWork.CommitAsync();
             }
             catch (Exception)
             {
+                // Rollback khi có lỗi xảy ra
                 await _unitOfWork.RollbackAsync();
                 throw;
             }
@@ -274,7 +323,7 @@ namespace MISA.WebFresher052023.Application.Service
             };
 
             // Kiểm tra điều kiện xóa
-            await _transferAssetManager.CheckDeleteTransferAssetAsync(transferAssetIds);
+            await _transferAssetManager.CheckDeleteManyTransferAssetAsync(transferAssetIds);
 
             // Lấy ra chứng từ cần xóa
             var entity = await _transferAssetRepository.GetAsync(transferAssetId);
@@ -291,24 +340,12 @@ namespace MISA.WebFresher052023.Application.Service
         /// <exception cref="UserException">Loại lệ người dùng</exception>
         public async Task DeleteManyAsync(List<Guid> transferAssetIds)
         {
-          
-            // Kiểm tra xem danh sách id rỗng
-            if (transferAssetIds.Count == 0)
-            {
-                throw new UserException(VietNamese.EmptyList);
-            }
 
             // Kiểm tra điều kiện xóa chứng từ
-            await _transferAssetManager.CheckDeleteTransferAssetAsync(transferAssetIds);
+            await _transferAssetManager.CheckDeleteManyTransferAssetAsync(transferAssetIds);
 
             // Lấy ra danh sách chứng từ cần xóa
             var fixedAssets = await _transferAssetRepository.FindManyAsync(transferAssetIds);
-
-            // Kiểm tra xem có chứng từ nào không tồn tại
-            if (fixedAssets.Count() != transferAssetIds.Count)
-            {
-                throw new UserException(VietNamese.NoDelete);
-            }
 
             // Lấy ra danh sách những tài sản điều chuyển cần xóa
             var transferAssetDetails = await _transferAssetDetailService.GetManyByTransferAssetIdsAsync(transferAssetIds);
@@ -319,27 +356,30 @@ namespace MISA.WebFresher052023.Application.Service
             // Mở Transaction
             await _unitOfWork.BeginTransactionAsync();
             try
-            {   
+            {
                 // Xóa nhiều tài sản điều chuyển
-                if (transferAssetDetails.Any())
+                if (transferAssetDetails != null && transferAssetDetails.Any())
                 {
                     await _transferAssetDetailService.DeleteManyAsync(transferAssetDetails.Select(transferAssetDetail => transferAssetDetail.TransferAssetDetailId).ToList());
                 }
-                 
+
                 // Xóa nhiều ban giao nhận
-                if (receivers.Any())
+                if (receivers != null && receivers.Any())
                 {
                     await _receiverService.DeleteManyAsync(receivers.Select(receiver => receiver.ReceiverId).ToList());
                 }
 
                 // Xóa nhiều chứng từ
-                await _transferAssetRepository.DeleteManyAsync(fixedAssets);
+                if (fixedAssets != null && fixedAssets.Any())
+                {
+                    await _transferAssetRepository.DeleteManyAsync(fixedAssets);
+                }
 
                 // Commit
                 await _unitOfWork.CommitAsync();
             }
             catch (Exception)
-            {   
+            {
                 // Rollback khi có lỗi xảy ra
                 await _unitOfWork.RollbackAsync();
                 throw;
